[FLINK-7176] [travis] Improve error handling
This closes #4330.

+++ tools/travis_mvn_watchdog.sh
@@ -272,7 +272,7 @@ check_shaded_artifacts() {
 		echo "=============================================================================="
 		echo "Detected $ASM asm dependencies in fat jar"
 		echo "=============================================================================="
-		exit 1
+		return 1
 	fi
 	 
 	GUAVA=`cat allClasses | grep '^com/google/common' | wc -l`
@@ -280,7 +280,7 @@ check_shaded_artifacts() {
 		echo "=============================================================================="
 		echo "Detected $GUAVA guava dependencies in fat jar"
 		echo "=============================================================================="
-		exit 1
+		return 1
 	fi
 
 	SNAPPY=`cat allClasses | grep '^org/xerial/snappy' | wc -l`
@@ -288,8 +288,9 @@ check_shaded_artifacts() {
 		echo "=============================================================================="
 		echo "Missing snappy dependencies in fat jar"
 		echo "=============================================================================="
-		exit 1
+		return 1
 	fi
+	return 0
 }
 
 # =============================================================================
@@ -346,37 +347,51 @@ echo "MVN exited with EXIT CODE: ${EXIT_CODE}."
 rm $MVN_PID
 rm $MVN_EXIT
 
-# Run tests
+# Run tests if compilation was successful
+if [ $EXIT_CODE == 0 ]; then
 
-echo "RUNNING '${MVN_TEST}'."
+	echo "RUNNING '${MVN_TEST}'."
 
-# Run $MVN_TEST and pipe output to $MVN_OUT for the watchdog. The PID is written to $MVN_PID to
-# allow the watchdog to kill $MVN if it is not producing any output anymore. $MVN_EXIT contains
-# the exit code. This is important for Travis' build life-cycle (success/failure).
-( $MVN_TEST & PID=$! ; echo $PID >&3 ; wait $PID ; echo $? >&4 ) 3>$MVN_PID 4>$MVN_EXIT | tee $MVN_OUT
+	# Run $MVN_TEST and pipe output to $MVN_OUT for the watchdog. The PID is written to $MVN_PID to
+	# allow the watchdog to kill $MVN if it is not producing any output anymore. $MVN_EXIT contains
+	# the exit code. This is important for Travis' build life-cycle (success/failure).
+	( $MVN_TEST & PID=$! ; echo $PID >&3 ; wait $PID ; echo $? >&4 ) 3>$MVN_PID 4>$MVN_EXIT | tee $MVN_OUT
 
-echo "Trying to KILL watchdog (${WD_PID})."
+	echo "Trying to KILL watchdog (${WD_PID})."
 
-# Make sure to kill the watchdog in any case after $MVN has completed
-( kill $WD_PID 2>&1 ) > /dev/null
+	# Make sure to kill the watchdog in any case after $MVN has completed
+	( kill $WD_PID 2>&1 ) > /dev/null
 
-EXIT_CODE=$(<$MVN_EXIT)
+	EXIT_CODE=$(<$MVN_EXIT)
 
-echo "MVN exited with EXIT CODE: ${EXIT_CODE}."
+	echo "MVN exited with EXIT CODE: ${EXIT_CODE}."
 
-rm $MVN_PID
-rm $MVN_EXIT
+	rm $MVN_PID
+	rm $MVN_EXIT
+else
+	echo "=============================================================================="
+	echo "Compilation failure detected, skipping test execution."
+	echo "=============================================================================="
+fi
 
 # Post
 
-# only misc builds flink-dist
+# only misc builds flink-dist and flink-yarn-tests
 case $TEST in
 	(misc)
-		check_shaded_artifacts
+		put_yarn_logs_to_artifacts
+
+		if [ $EXIT_CODE == 0 ]; then
+			check_shaded_artifacts
+			EXIT_CODE=$?
+		else
+			echo "=============================================================================="
+			echo "Compilation/test failure detected, skipping shaded dependency check."
+			echo "=============================================================================="
+		fi
+
 	;;
 esac
-
-put_yarn_logs_to_artifacts
 
 upload_artifacts_s3
